Update récupération des données pipeline
